Improving client project structure
