+++ client/.env
@@ -1,0 +1,2 @@
+#-----------------------------API-------------------
+NEXT_PUBLIC_API_URL=http://localhost:5000/api
