+++ client/src/app/(auth)/layout.tsx
@@ -1,14 +1,3 @@
-import { ReactNode } from "react";
-import styles from "./AuthLayout.module.scss";
+import { AuthLayout } from "@/layouts/Auth";
 
-type AuthLayoutProps = {
-  children: ReactNode;
-};
-
-export default function AuthLayout({ children }: AuthLayoutProps) {
-  return (
-    <div className={styles.root}>
-      <div className={styles.container}>{children}</div>
-    </div>
-  );
-}
+export default AuthLayout;
