+++ client/src/app/layout.tsx
@@ -1,32 +1,11 @@
 import type { Metadata } from "next";
 import "../styles/_normalize.scss";
 import "../styles/_variables.scss";
-import { Montserrat } from "next/font/google";
-import { ReactQueryProvider } from "@/providers";
-
-const montserrat = Montserrat({
-  subsets: ["latin"],
-  weight: ["100", "200", "300", "400", "500"],
-  variable: "--primary-font",
-});
+import { RootLayout } from "@/layouts/Root";
 
 export const metadata: Metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
 };
 
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body>
-        <ReactQueryProvider>
-          <main>{children}</main>
-        </ReactQueryProvider>
-      </body>
-    </html>
-  );
-}
+export default RootLayout;
