+++ client/src/layouts/Auth/AuthLayout.tsx
@@ -1,0 +1,14 @@
+import { ReactNode } from "react";
+import styles from "./AuthLayout.module.scss";
+
+type AuthLayoutProps = {
+  children: ReactNode;
+};
+
+export const AuthLayout: React.FC<AuthLayoutProps> = ({ children }) => {
+  return (
+    <div className={styles.root}>
+      <div className={styles.container}>{children}</div>
+    </div>
+  );
+};
