+++ client/src/layouts/Root/index.ts
@@ -1,0 +1,1 @@
+export * from "./RootLayout";
