+++ client/src/pages/notFound/NotFoundPage.tsx
@@ -1,0 +1,4 @@
+type NotFoundPageProps = {};
+export const NotFoundPage: React.FC<NotFoundPageProps> = () => {
+  return <>Not found</>;
+};
