+++ client/src/pages/notFound/index.ts
@@ -1,0 +1,1 @@
+export * from "./NotFoundPage";
